% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2html.R
\name{dataframe2html}
\alias{dataframe2html}
\title{Convert dataframe to a table within an HTML page}
\usage{
dataframe2html(data)
}
\arguments{
\item{data}{A dataframe containing columns and rows of data.}
}
\value{
an HTML object containing the full HTML code for the webpage. The function
also saves the HTML file to the working directory.
}
\description{
Function converts a dataframe into an HTML table embedded within a full
HTML file. The output file contains a summary showing the total number of table rows.
The HTML file also includes JavaScript functions to support URL suffix searching via
appending '?s=\link{query}' to the final URL. A JavaScript function filters the table and
displays matching rows (based on an exact match search across all columns), updating
the disaplyed row totals in the summary line.
}
